crypto-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
cryptography==41.0.7
argon2-cffi==23.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0